Mining profile
This is my current hash rate profile - the dips are at night and my fan drops below 3000RPM (zzzzzzz) but that's a cost - I need to underclock to 200MHz to achieve that. In the morning I overclock to 375 for 4 hours (you can see the peak) then down to 350 until bed time.